project_root/requirements.txt
python-telegram-bot==20.8
openai==1.12.0
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
SQLAlchemy==2.0.27
aiosqlite==0.20.0 